# flights-api/src/main/resources/swagger.yaml
# Se incluye este .yaml para poder acceder a la documentación en Swagger que se encuentra en el README.md mediante un enlace
openapi: 3.0.1
info:
  title: Flights API
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /flights/update/{id}:
    put:
      tags:
        - flight-controller
      operationId: updateFlight
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Flight'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Flight'
  /companies/update/{id}:
    put:
      tags:
        - company-controller
      operationId: updateCompany
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Company'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Company'
  /flights/create/{companyId}:
    post:
      tags:
        - flight-controller
      operationId: createFlight
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Flight'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Flight'
  /companies/create:
    post:
      tags:
        - company-controller
      operationId: createCompany
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Company'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Company'
  /flights:
    get:
      tags:
        - flight-controller
      operationId: getAllFlights
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FlightDto'
  /flights/{id}:
    get:
      tags:
        - flight-controller
      operationId: getFlightById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Flight'
  /flights/origin:
    get:
      tags:
        - flight-controller
      operationId: getFlightsByOrigin
      parameters:
        - name: origin
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FlightDto'
  /flights/offers:
    get:
      tags:
        - flight-controller
      operationId: getOffers
      parameters:
        - name: offerPrice
          in: query
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FlightDto'
  /flights/location:
    get:
      tags:
        - flight-controller
      operationId: getFlightsByLocations
      parameters:
        - name: origin
          in: query
          required: true
          schema:
            type: string
        - name: destiny
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FlightDto'
  /flights/company:
    get:
      tags:
        - flight-controller
      operationId: getFlightsByCompany
      parameters:
        - name: companyName
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FlightDto'
  /companies:
    get:
      tags:
        - company-controller
      operationId: getAllCompanies
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Company'
  /companies/{id}:
    get:
      tags:
        - company-controller
      operationId: getCompanyById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Company'
  /companies/name:
    get:
      tags:
        - company-controller
      operationId: getCompanyByName
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Company'
  /flights/delete/{id}:
    delete:
      tags:
        - flight-controller
      operationId: deleteFlight
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseDto'
  /companies/delete/{id}:
    delete:
      tags:
        - company-controller
      operationId: deleteCompany
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseDto'
components:
  schemas:
    Company:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        page:
          type: string
        banner:
          type: string
    Flight:
      type: object
      properties:
        id:
          type: integer
          format: int64
        origin:
          type: string
        destiny:
          type: string
        departureTime:
          type: string
          format: date-time
        arrivingTime:
          type: string
          format: date-time
        price:
          type: number
          format: double
        frequency:
          type: string
        company:
          $ref: '#/components/schemas/Company'
    FlightDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        origin:
          type: string
        destiny:
          type: string
        departureTime:
          type: string
          format: date-time
        arrivingTime:
          type: string
          format: date-time
        convertedPrice:
          type: number
          format: double
        frequency:
          type: string
        company:
          $ref: '#/components/schemas/Company'
    ResponseDto:
      type: object
      properties:
        message:
          type: string